Added attribute types man ui

+++ std-extensions/src/main/java/pl/edu/icm/unity/stdext/attr/JpegImageAttributeSyntax.java
@@ -172,7 +172,7 @@ public class JpegImageAttributeSyntax implements AttributeValueSyntax<BufferedIm
 		return maxWidth;
 	}
 
-	public long getMaxHeight()
+	public int getMaxHeight()
 	{
 		return maxHeight;
 	}
